+++ src/extension/features/accounts/hide-memos/index.js
@@ -1,8 +1,40 @@
 import * as React from 'react';
 import { componentAppend } from 'toolkit/extension/utils/react';
-import { HideMemosButton } from './components/hide-memos-button';
 import { Feature } from 'toolkit/extension/features/feature';
-import { getToolkitStorageKey, setToolkitStorageKey } from 'toolkit/extension/utils/toolkit';
+import { l10n, getToolkitStorageKey, setToolkitStorageKey } from 'toolkit/extension/utils/toolkit';
+import * as PropTypes from 'prop-types';
+import { controllerLookup } from 'toolkit/extension/utils/ember';
+
+export const HideMemosButton = ({ toggleHiddenState }) => {
+  const notHidden = getToolkitStorageKey('hide-memos', true);
+  const label = notHidden ? 'is-checked' : 'not-checked';
+
+  const toggleHidden = () => {
+    toggleHiddenState(!notHidden);
+    controllerLookup('application').send('closeModal');
+  };
+
+  return (
+    <div className="modal-account-view-menu">
+      <button
+        onClick={toggleHidden}
+        className={`${label}`}
+        aria-label="Check Hide Memo Column"
+        role="checkbox"
+      >
+        <div className={`flaticon stroke ynab-checkbox-button-square ${label}`}></div>
+      </button>
+      <label className="label-checkbox">
+        {` `}
+        {l10n('toolkit.hideMemoColumn', 'Show Memo Column')}
+      </label>
+    </div>
+  );
+};
+
+HideMemosButton.propTypes = {
+  toggleHiddenState: PropTypes.func.isRequired,
+};
 
 export class HideMemos extends Feature {
   injectCSS() {
